[Non-executable version] Clean up database structure
