+++ app/auth.py
@@ -11,12 +11,8 @@ from fastapi_users.db import SQLAlchemyUserDatabase
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from .config import settings
-from .db import get_session
-from .model import User
-
-
-async def get_user_db(session: AsyncSession = Depends(get_session)):
-    yield SQLAlchemyUserDatabase(session, User)
+from app.database.tools import get_session
+from app.model.user import User
 
 
 class UserManager(UUIDIDMixin, BaseUserManager[User, uuid.UUID]):
@@ -24,23 +20,23 @@ class UserManager(UUIDIDMixin, BaseUserManager[User, uuid.UUID]):
     verification_token_secret = settings.AUTH_SECRET
 
 
+async def get_user_db(session: AsyncSession = Depends(get_session)):
+    yield SQLAlchemyUserDatabase(session, User)
+
+
 async def get_user_manager(user_db=Depends(get_user_db)):
     yield UserManager(user_db)
-
-
-bearer_transport = BearerTransport(tokenUrl='/api/v1/auth/jwt/login')
 
 
 def get_jwt_strategy() -> JWTStrategy:
     return JWTStrategy(secret=settings.AUTH_SECRET, lifetime_seconds=3600)
 
 
+bearer_transport = BearerTransport(tokenUrl='/api/v1/auth/jwt/login')
 auth_backend = AuthenticationBackend(
     name='jwt',
     transport=bearer_transport,
     get_strategy=get_jwt_strategy,
 )
-
 fastapi_users = FastAPIUsers[User, uuid.UUID](get_user_manager, [auth_backend])
-
 current_active_user = fastapi_users.current_user(active=True)
