+++ app/database/tools.py
@@ -1,9 +1,9 @@
 from typing import AsyncIterable
 
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
-from sqlalchemy.orm import DeclarativeBase, sessionmaker
+from sqlalchemy.orm import sessionmaker
 
-from .config import settings
+from app.config import settings
 
 
 engine = create_async_engine(settings.DATABASE_URL, echo=settings.DEBUG)
@@ -12,10 +12,6 @@ async_session = sessionmaker(
     class_=AsyncSession,
     expire_on_commit=False,
 )
-
-
-class Base(DeclarativeBase):
-    pass
 
 
 async def get_session() -> AsyncIterable[AsyncSession]:
