+++ app/model/user.py
@@ -1,0 +1,7 @@
+from fastapi_users.db import SQLAlchemyBaseUserTableUUID
+
+from app.database.base import Base
+
+
+class User(SQLAlchemyBaseUserTableUUID, Base):
+    pass
